requirements.txt
fastapi
uvicorn
tqdm
pandas
python-multipart
python-dotenv
sentence-transformers
qdrant-client